# Conflicts:
#	README.md
#	ultralytics/yolo/utils/__init__.py

+++ ultralytics/yolo/utils/checks.py
@@ -19,8 +19,8 @@ import requests
 import torch
 from matplotlib import font_manager
 
-from ultralytics.yolo.utils import (AUTOINSTALL, LOGGER, ONLINE, RANK, ROOT, USER_CONFIG_DIR, TryExcept, clean_url,
-                                    colorstr, downloads, emojis, is_colab, is_docker, is_kaggle, is_online,
+from ultralytics.yolo.utils import (AUTOINSTALL, LOGGER, ONLINE, ROOT, USER_CONFIG_DIR, TryExcept, clean_url, colorstr,
+                                    downloads, emojis, is_colab, is_docker, is_jupyter, is_kaggle, is_online,
                                     is_pip_package, url2file)
 
 
@@ -47,7 +47,7 @@ def check_imgsz(imgsz, stride=32, min_dim=1, max_dim=2, floor=0):
     stride, update it to the nearest multiple of the stride that is greater than or equal to the given floor value.
 
     Args:
-        imgsz (int) or (cList[int]): Image size.
+        imgsz (int | cList[int]): Image size.
         stride (int): Stride value.
         min_dim (int): Minimum number of dimensions.
         floor (int): Minimum allowed value for image size.
@@ -164,26 +164,23 @@ def check_font(font='Arial.ttf'):
     Returns:
         file (Path): Resolved font file path.
     """
-    from ultralytics.yolo.utils.torch_utils import torch_distributed_zero_first
+    name = Path(font).name
 
-    with torch_distributed_zero_first(RANK):
-        name = Path(font).name
+    # Check USER_CONFIG_DIR
+    file = USER_CONFIG_DIR / name
+    if file.exists():
+        return file
 
-        # Check USER_CONFIG_DIR
-        file = USER_CONFIG_DIR / name
-        if file.exists():
-            return file
+    # Check system fonts
+    matches = [s for s in font_manager.findSystemFonts() if font in s]
+    if any(matches):
+        return matches[0]
 
-        # Check system fonts
-        matches = [s for s in font_manager.findSystemFonts() if font in s]
-        if any(matches):
-            return matches[0]
-
-        # Download to USER_CONFIG_DIR if missing
-        url = f'https://ultralytics.com/assets/{name}'
-        if downloads.is_url(url):
-            downloads.safe_download(url=url, file=file)
-            return file
+    # Download to USER_CONFIG_DIR if missing
+    url = f'https://ultralytics.com/assets/{name}'
+    if downloads.is_url(url):
+        downloads.safe_download(url=url, file=file)
+        return file
 
 
 def check_python(minimum: str = '3.7.0') -> bool:
@@ -328,8 +325,11 @@ def check_yolo(verbose=True, device=''):
     """Return a human-readable YOLO software and hardware summary."""
     from ultralytics.yolo.utils.torch_utils import select_device
 
-    if is_colab():
-        shutil.rmtree('sample_data', ignore_errors=True)  # remove colab /sample_data directory
+    if is_jupyter():
+        if check_requirements('wandb', install=False):
+            os.system('pip uninstall -y wandb')  # uninstall wandb: unwanted account creation prompt with infinite hang
+        if is_colab():
+            shutil.rmtree('sample_data', ignore_errors=True)  # remove colab /sample_data directory
 
     if verbose:
         # System info
